Responding to Anna Maria's Feedback
Page 7 - Is there a reason that the sites didn't respond?  Did they decline to respond?  Did they not have an issue or an interest?  The wording is a bit add without any real explanation.

Page 8 – Total power load greater than 10MW? We should clarify that this total HPC load includes computational and ancillary support load (mechanical, etc.)

Page 8 – The variability at the site is not due to the HPC  being > than 10MW, it is the total site load which dictates the variability.  This is not clear in the paragraph above Table 1.

Page 9 – The statement "More than 1/2 of the respondents are not interested in shedding or shifting….." is not entirely accurate.  The paragraph later contradicts this by saying that shifting may be of an interest which could be true or advantageous in order to keep the load stable in times of grid nuances.

Page 9 – The quotation from LLNL is a bit odd.  It should be changed to LLNL is " responding to requests for additional data hourly, weekly and monthly for HPC loads and is also working to develop an automated capability to share data between them and us which would provide automated additional detailed forecasting and ultimately real time data".

Anna Maria