Added PUE results
Moved discussion from conclusions to questionnaire.

It is clear
from the HPC operation that the largest opportunity for load management exists in the IT equipment. However, the previous work suggests that the opportunity
in HVAC loads may also be an immediate opportunity. The opportunity
depends on the ratio of cooling loads to the IT equipment load (e.g., PUE). Considering
that the lighting loads are a small percentage of overall HPC load, their
participation alongside other strategies (IT equipment and HVAC) may prove
beneficial. The advancement in technologies and vendor recognition to provide
more dynamic power management capabilities in the IT equipment offers larger
opportunity in consolidated IT and cooling load reductions.